Version 0.47247

+++ butm.sh
@@ -7,7 +7,7 @@
 MUPR=$(readlink -f $0); # Mutterprogramm
 . ${MUPR%/*}/bul1.sh # LINEINS=linux1, buhost aus hostname festlegen
 [ "$buhost"/ = "$LINEINS"/ ]&&ZL=||QL=$LINEINS;
-. ${MUPR%/*}/bugem.sh # commandline-Parameter, $ZL aus commandline, $qssh, $zssh festlegen
+. ${MUPR%/*}/bugem.sh # commandline-Parameter, $ZL aus commandline, in obalt und kopiermt $qssh, $zssh festlegen
 [ "$buhost"/ != "$LINEINS"/ -a "$ZL" ]&&{ printf "Ziel \"$blau$ZL$reset\" wird zurückgesetzt.\n"; ZL=;}
 [ "$buhost"/ = "$LINEINS"/ -a -z "$ZL" ]&&{ printf "${rot}Kein Ziel angegeben. Breche ab$reset.\n";exit;}
 # [ "$QL" ]&& pruefpc "$QL";
@@ -15,12 +15,12 @@ MUPR=$(readlink -f $0); # Mutterprogramm
 wirt=${QL:-$buhost};
 . ${MUPR%/*}/virtnamen.sh # legt aus $wirt fest: $gpc, $gast, $tush
 g1=$gast;
-l1gpc=$gpc; # Gast-PC von Linux1
+l1gpc=$gpc; # Gast-PC von Linux1, also virtwin
 wirt=${ZL:-$buhost};
 . ${MUPR%/*}/virtnamen.sh # legt aus $wirt fest: $gpc, $gast, $tush
 rgpc=$gpc; # Gast-PC des Reserveservers
 # hier wir tush nicht nicht wirt festgelegt, wie in virtnamen.sh, sondern nach buhost:
-case $buhost in $LINEINS)tush="sh -c ";;*)tush="ssh $buhost ";;esac
+case $buhost in $LINEINS)tush="sh -c ";zish="ssh $buhost ";;*)tush="ssh $buhost ";zish="sh -c ";;esac
 
 ot=opt/turbomed;
 otP=/$ot/PraxisDB;
@@ -28,11 +28,11 @@ otr=/$ot/PraxisDB-res;
 if eval "$tush 'test -d $otP'"; then # wenn es auf linux1 /opt/turbomed/PraxisDB gibt, 
     obvirt=;                                   # also nicht die virtuelle Installation verwendet wird
     Pr=PraxisDB;
-    ausf "$zssh '[ -d $otr -a ! -d $otP ]&& mv $otr $otP'" $blau; # umgekehrt
+    ausf "$zish '[ -d $otr -a ! -d $otP ]&& mv $otr $otP'" $blau; # umgekehrt
 else
     obvirt=1; 
     Pr=PraxisDB-res;
-    ausf "$zssh '[ -d $otP -a ! -d $otr ]&& mv $otP $otr'" $blau; # dann ggf. auf dem Zielrechner die linux-Datenbank umbenennen
+    ausf "$zish '[ -d $otP -a ! -d $otr ]&& mv $otP $otr'" $blau; # dann ggf. auf dem Zielrechner die linux-Datenbank umbenennen
 fi;
 for iru in 1 2; do
   if test $iru = 1; then
@@ -48,25 +48,31 @@ for iru in 1 2; do
     uZL=$ZL;
     ZL=; # dann werden die cifs-Laufwerke verwendet
     [ "$obkill" ]&&{ if mountpoint -q /$ur||[ $iru = 1 ];then 
-      mv /$ur/lauf /$ur/lau&&sleep 1m;
-      ausf "VBoxManage controlvm $g1 poweroff" $blau;
-      ausf "VBoxManage startvm $g1 --type headless" $blau;
+      if ! ssh administrator@$l1gpc cmd /c "(>>c:\turbomed\StammDB\objects.idx (call ) )&&exit||exit /b 1" 2>/dev/nul; then
+       ausf "$tush 'mv /$ot/lauf /$ot/lau '&&sleep 1m";
+      fi;
+      if ! ssh administrator@$l1gpc cmd /c "(>>c:\turbomed\StammDB\objects.idx (call ) )&&exit||exit /b 1" 2>/dev/nul; then
+        ausf "VBoxManage controlvm $g1 poweroff" $blau;
+        ausf "VBoxManage startvm $g1 --type headless" $blau;
+      fi;
     fi;};# dann killt der windows-task "Turbomed töten" turbomed
 
   fi;
   [ "$verb" ]&&printf "tush: ${blau}$obsh$reset\n";
   [ "$verb" ]&&printf "obvirt: ${blau}$obvirt$reset\n";
   [ "$obforce" ]&&testdat=||testdat=$Pr/objects.dat;
-  kopiermt "$ur/" "$hin" "" "$OBDEL" "$testdat" "1800" 1; # ohne --iconv
+  if ssh administrator@$l1gpc cmd /c "(>>c:\turbomed\StammDB\objects.idx (call ) )&&exit||exit /b 1" 2>/dev/nul; then
+   kopiermt "$ur/" "$hin" "" "$OBDEL" "$testdat" "1800" 1; # ohne --iconv
+  fi;
 done;
-[ "$obkill" -a "$obvirt" ]&&{ mv /$ur/lau /$ur/lauf||touch /$ur/lauf;} # zurückbenennen, damit Turbomed wieder starten kann
+[ "$obkill" -a "$obvirt" ]&&{ mv /$ot/lau /$ot/lauf||touch /$ot/lauf;} # zurückbenennen, damit Turbomed wieder starten kann
 ZL=$altZL;
 Dt=DATA; 
 Pt=Patientendokumente;
-for zug in "$qssh" "$zssh"; do
+for zug in "$tush" "$zish"; do
   ausf "$zug 'mountpoint -q /$Dt 2>/dev/null||mount /$Dt'";
 done;
-ausf "$qssh 'mountpoint -q /${Dt} 2>/dev/null'&&$zssh 'mountpoint -q /${Dt} 2>/dev/null'"
+ausf "$tush 'mountpoint -q /${Dt} 2>/dev/null'&&$zish 'mountpoint -q /${Dt} 2>/dev/null'"
 if [ "$ret"/ = 0/ ]; then
  QL=$uQL;
  ZL=$uZL;
